requirements.txt
fastapi[standard]==0.115.9
orjson==3.10.15
pydantic-settings==2.8.1
python-dotenv==1.0.1
python-logstash==0.4.8